Re-organized server-specific files a bit.
